backend/requirements.txt
appdirs==1.4.4
attrs==21.2.0
blis==0.7.5
boto3==1.17.32
botocore==1.20.32
catalogue==2.0.6
certifi==2020.12.5
cffi==1.15.0
cfgv==3.2.0
chardet==4.0.0
click==8.0.3
cryptography==35.0.0
cymem==2.0.6
deepdiff==5.7.0
distlib==0.3.1
#en-core-web-sm==3.1.0
falcon==2.0.0
falcon-pagination==0.0.1
falcon-params-verifier==0.1.0
filelock==3.0.12
gunicorn==20.0.4
identify==2.0.0
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.1.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
more-itertools==8.10.0
murmurhash==1.0.6
nodeenv==1.5.0
numpy==1.21.2
ordered-set==4.0.2
packaging==21.0
pandas==0.25.3
pathy==0.6.0
pluggy==0.13.1
pre-commit==2.10.1
preshed==3.0.6
psycopg2-binary==2.8.4
py==1.10.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.3
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.7
six==1.15.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.2
thinc==8.0.11
toml==0.10.2
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.7.4.3
urllib3==1.26.5
virtualenv==20.4.2
wasabi==0.8.2
wcwidth==0.2.5
zipp==3.4.0
